part 4.21

+++ part4/blog/controllers/blog.js
@@ -36,12 +36,24 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    const result = await Blog.findByIdAndDelete(request.params.id)
-    if (result) {
-        response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if (blog) {
+        if (blog.user.toString() === decodedToken.id.toString()) {
+            await Blog.findByIdAndDelete(request.params.id)
+            response.status(204).end()
+        } else {
+            response.status(401).end()
+        }
     } else {
         response.status(404).end()
     }
+
+
 })
 
 blogRouter.put('/:id', async (request, response) => {
